print debug messages to stdout (#163)
